widen mujs instruction size

+++ mujs/jsi.h
@@ -76,7 +76,7 @@ typedef struct js_StackTrace js_StackTrace;
 #define JS_ASTLIMIT 100		/* max nested expressions */
 
 /* instruction size -- change to int if you get integer overflow syntax errors */
-typedef unsigned short js_Instruction;
+typedef unsigned js_Instruction;
 
 /* String interning */
